move the poly multiplications to solvers, update everything
