+++ docs/source/geometry.epipolar.rst
@@ -52,9 +52,7 @@ Numeric
 -------
 
 .. autofunction:: cross_product_matrix
-.. autofunction:: o1
-.. autofunction:: o2
-.. autofunction:: det_to_poly
+
 
 Triangulation
 -------------
