+++ docs/source/geometry.solvers.rst
@@ -10,3 +10,7 @@ Polynomial Solvers
 
 .. autofunction:: solve_quadratic
 .. autofunction:: solve_cubic
+.. autofunction:: multiply_deg_one_poly
+.. autofunction:: multiply_deg_two_one_poly
+.. autofunction:: determinant_to_polynomial
+
