+++ kornia/geometry/solvers/__init__.py
@@ -1,3 +1,3 @@
-from .polynomial_solver import solve_cubic, solve_quadratic
+from .polynomial_solver import solve_cubic, solve_quadratic, multiply_deg_one_poly, multiply_deg_two_one_poly, determinant_to_polynomial
 
-__all__ = ["solve_quadratic", "solve_cubic"]
+__all__ = ["solve_quadratic", "solve_cubic", 'multiply_deg_one_poly', 'multiply_deg_two_one_poly', 'determinant_to_polynomial']
